Reduce card match checking timeout
Card match checking timeout was reduced as clicking cards too quickly messes the
game up a bit. This is not a solution though, more like a workaround to reduce
the possibility for such cases to happen. This also makes the game a bit harder
as opened non-matching cards are visible for shorter time.